Chunk refactoring
Changed Chunk to ChunkSection and ChunkColumn to Chunk to avoid confusion
Moved the palette data from Chunk, as it is specific of ChunkSection
