+++ elytra-sdk/src/main/kotlin/io/elytra/sdk/world/ElytraChunkSection.kt
@@ -1,0 +1,38 @@
+package io.elytra.sdk.world
+
+import io.elytra.api.world.Chunk
+
+class ElytraChunk() : Chunk {
+
+    override fun createColumns() {
+        TODO("not implemented") // To change body of created functions use File | Settings | File Templates.
+    }
+
+    override fun getChunkAt(y: Int) {
+        TODO("not implemented") // To change body of created functions use File | Settings | File Templates.
+    }
+
+    override fun getBlockAt(x: Int, y: Int, z: Int) {
+        TODO("not implemented") // To change body of created functions use File | Settings | File Templates.
+    }
+
+    override fun getSkyLight(x: Int, y: Int, z: Int) {
+        TODO("not implemented") // To change body of created functions use File | Settings | File Templates.
+    }
+
+    override fun getBlockLight(x: Int, y: Int, z: Int) {
+        TODO("not implemented") // To change body of created functions use File | Settings | File Templates.
+    }
+
+    override fun setSkyLight(x: Int, y: Int, z: Int, skyLight: Int) {
+        TODO("not implemented") // To change body of created functions use File | Settings | File Templates.
+    }
+
+    override fun setBlockLight(x: Int, y: Int, z: Int, blockLight: Int) {
+        TODO("not implemented") // To change body of created functions use File | Settings | File Templates.
+    }
+
+    override fun getBiome(x: Int, z: Int) {
+        TODO("not implemented") // To change body of created functions use File | Settings | File Templates.
+    }
+}
